travel-suanfa/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
psycopg2-binary==2.9.9
numpy==1.26.2
scipy==1.11.4
python-multipart==0.0.6
httpx==0.25.2